Delete file handler_token_issue.go
